+++ src/catppuccinTheme.latte.ts
@@ -1,3 +1,3 @@
 import { createCatppuccinPlugin } from '@catppuccin/daisyui'
 
-export default createCatppuccinPlugin('latte', {}, { default: true })
+export default createCatppuccinPlugin('latte')